// src/types/product.ts
export type Product = {
  title: string;
  reviews: number;
  price: number;
  discountedPrice: number;
  id: number | string;
  imgs?: {
    thumbnails: string[];
    previews: string[];
  };
  categories?: string[];
  shortDescription?: any;
  presentation?: string;
  slug?: string;
  isHeroCarousel?: boolean;
  isHeroOferta?: boolean;
  isPromoSection?: boolean;
};